refactor(convert): split request, xml shapes, schema store, toolmap
