+++ backend/internal/convert/CONTRACT.md
@@ -26,7 +26,7 @@ EVERYTHING else — `config`, `pool`, `upstream`, `server`, `registry`, `dashboa
 ## Critical invariants
 
 - Pure: the `*Opts` variants never read the process environment, touch the filesystem, or make network calls; the only external inputs are passed in explicitly (`Options`, a `ReasoningLookupFn`, an Accumulator's per-call config). `DefaultOptions` is a deprecated compatibility shim that still reads `COMPRESS_PROMPT` / `CACHE_CONTROL_INJECTION` / `REASONING_IN_CONTENT` from the environment until callers pass config-derived options (issue #277).
-- `end_turn` injection: the proxy injects the `end_turn` tool definition to pass Codebuff's `foreign_toolset` validation (schemacache.go). Downstream parity: it MUST be stripped before emitting to clients (`StripEndTurnToolCalls`); both the OpenAI and Anthropic relays call it. An existing `end_turn` must never be duplicated.
+- `end_turn` injection: the proxy injects the `end_turn` tool definition to pass Codebuff's `foreign_toolset` validation (schemacache_endturn.go). Downstream parity: it MUST be stripped before emitting to clients (`StripEndTurnToolCalls`); both the OpenAI and Anthropic relays call it. An existing `end_turn` must never be duplicated.
 - Schema normalization: `$ref` inlining with cycle guard, depth cap (12 levels), per-request node budget, enum cross-type dedupe; the schema cache returns maps that callers must not mutate (mutation-safety pinned by test).
 - Stream sanitization: whitelist keys, drop empty-choice chunks, drop enrichment keys (`service_tier`, `obfuscation`, `moderation`), correct `finish_reason`/`[DONE]` handling; canonical chunks take a fast path.
 - XML extractor: pooled buffers, bounded partial-hold (repeated `<` cannot re-hold forever), blocks close on `</tool_call>`; extraction of `end_turn`-shaped blocks must never surface to clients.
